finish truck database change

+++ src/main/java/mapper/PackageMapper.java
@@ -11,7 +11,7 @@ public interface PackageMapper {
     List<Package> findAll();
 
     @Select("SELECT * FROM package WHERE packageId = #{packageId}")
-    Package findByPackageId(@Param("packageId") Integer packageId);
+    Package findByPackageId(@Param("packageId") Long packageId);
     @Select("SELECT * FROM package WHERE userId = #{userId}")
 
     List<Package> findByUserId(@Param("userId") Integer userId);
